Make the flow bar actually play/pause/rewind/stop
Addresses
---------
Need to have play/pause/rewind/stop capabilities.

Why
---
* I have plans to replace the button text with icons but they're not working yet.
* I have chosen a combination of DOM API and Coffeescript because I'm not that
  familiar with jquery and dont want to learn it in this short period.
* Therefore the code for handleRewind gets a bit hairy. If JS were completely
  synchronous, there'd be 2 delay loops in its body - one to delay the undos and
  one for the redos. However, since JS is async, I've implemented the delays
  between the undo steps via setInterval, which by default uses window as its
  context, so we need to bind the right function reference (see line 77). doUndo
  calls doRedo once its 15 rounds of undoing are met, which allows the user to
  see a replay of the last 15 edit actions, thereby hopefully jogging the grey
  cells :)
* Atom limitation: There's no global undo - only per file. Thus you cannot see
  rewind/replay past the last undo on the file. If this tool integrates with
  groktools/lilbro, then we might be able to do that.

What
----
added ability to disable buttons at start
changed textbook to atom-text-editor and set placeholder text correctly
figured out the right way to bind this in js, not sure if this is right in atom,
possibly via events and subscriptions.
handled the state changes for the start/resume button,the task name textbox
and stop button, relatively easy
handled the state changes for the rewind button which required quite a bit of
work since the only way to add a delay in js is using set interval, which is
async. the handlers need to access the button text to keep it updated, so that
required a lot of googling to correctly bind 'this'. line 77 is that way.
also note that doUndo and doRedo cascade. This should probably be replaced with
promises.