Add icons to the flow bar and improve replay
Addresses
---------
Better aesthetics and functionality

Why
---
The previous commit worked mostly but it could be better. This commit intends to
fix that by making the buttons look more like buttons ie with icons. This needed
more work than I expected. First, I had to figure out how to include font-awesome
into the codebase and looked at other atom packages for how they did it, links
to appropriate atom discussions are in the code.

To make the icon glyphs work, the buttons had to be made span elements.

Then making the buttons work as buttons when they're really span elements was
challenging and required creating the @enableButton, @disableButton, @switchIcons
functions.

I also created the @cmdMap so that the display names on the buttons are
disconnected from the icon names.

Finally, I switched the replay and stop buttons around because I felt that having
the stop to the right of replay gave the impression that the replay can be stopped,
while the intent is for the dev to press stop to stop working on the task.