Add a logo to the flow bar
Addresses
---------
Need to highlight the flow bar in the UI

Why
---
I used the fa-space-shuttle icon since that was the closest to the  concept of
"flow" that the font awesome library had.

Introducing the logo meant converting the old "Flozone" span to an <i> element,
so I added the @addLogo method for that.

Note: coffeescript is not ruby. @addLogo is not a call, @addLogo() is.