Fix issue: stop doesnt reset the start button
Addresses
---------

Why
---
handleStop had two issues:
1. It didnt really handle the paused -> stop transition
2. It didnt reset the start button correctly.

I fixed the first issues by adding paused to the when block, but wrongly thought
that the reset of the start button was not happening because the old css selector
for font-awesome was not cleared correctly. So I stored the old selector in the
button's dataset attribute (ie custom html data- attr) and used that to clear
the css correctly. However, I later realized that the real problem was that i was
not setting the button text back to "start" and that was the real problem.

However, the move to using the data element to store the prev selector had 2 uses:
1. It actually fixed a subtle problem with the icons not being right during
   transactions.
2. It was better than specifying the previous icon every time.